first attempt at dir structure cleanup
