+++ include/os/freebsd.h
@@ -1,0 +1,6 @@
+/* freebsd.h - A secure memory allocator
+ * Copyright 2022 - chris.rohlf@gmail.com */
+
+#pragma once
+
+#define MAP_HUGETLB MAP_ALIGNED_SUPER
