+++ src/iso_alloc_interfaces.c
@@ -3,6 +3,7 @@
 
 #include "iso_alloc.h"
 #include "iso_alloc_internal.h"
+#include "iso_alloc_ds.h"
 
 #if MEMCPY_SANITY
 #include "iso_alloc_sanity.h"
@@ -61,11 +62,7 @@ EXTERNAL_API NO_DISCARD REALLOC_SIZE ASSUME_ALIGNED void *iso_realloc(void *p, s
     }
 
     if(p != NULL) {
-#if MEMCPY_SANITY
-        __iso_memcpy(r, p, size);
-#else
-        __builtin_memcpy(r, p, size);
-#endif
+        _iso_alloc_memcpy(r, p, size);
     }
 
 #if PERM_FREE_REALLOC
@@ -108,11 +105,7 @@ EXTERNAL_API NO_DISCARD ASSUME_ALIGNED char *iso_strdup_from_zone(iso_alloc_zone
         return NULL;
     }
 
-#if MEMCPY_SANITY
-    __iso_memcpy(p, str, size);
-#else
-    __builtin_memcpy(p, str, size);
-#endif
+    _iso_alloc_memcpy(p, str, size);
     return p;
 }
 
@@ -138,18 +131,10 @@ EXTERNAL_API NO_DISCARD ASSUME_ALIGNED char *iso_strndup_from_zone(iso_alloc_zon
     }
 
     if(s_size > n) {
-#if MEMCPY_SANITY
-        __iso_memcpy(p, str, n);
-#else
-        __builtin_memcpy(p, str, n);
-#endif
+        _iso_alloc_memcpy(p, str, n);
         p[n - 1] = '\0';
     } else {
-#if MEMCPY_SANITY
-        __iso_memcpy(p, str, s_size);
-#else
-        __builtin_memcpy(p, str, s_size);
-#endif
+        _iso_alloc_memcpy(p, str, s_size);
     }
 
     return p;
@@ -227,7 +212,8 @@ EXTERNAL_API int32_t iso_alloc_name_zone(iso_alloc_zone_handle *zone, char *name
         UNMASK_ZONE_HANDLE(zone);
     }
 
-    return name_zone(zone, name);
+    iso_alloc_zone_t *_zone = (iso_alloc_zone_t *) zone;
+    return name_mapping(_zone->user_pages_start, ZONE_USER_SIZE, name);
 }
 
 EXTERNAL_API void iso_alloc_protect_root() {
