+++ src/iso_alloc_util.c
@@ -126,14 +126,6 @@ INTERNAL_HIDDEN void mprotect_pages(void *p, size_t size, int32_t protection) {
     }
 }
 
-INTERNAL_HIDDEN int32_t name_zone(iso_alloc_zone_t *zone, char *name) {
-#if NAMED_MAPPINGS && __ANDROID__
-    return name_mapping(zone->user_pages_start, ZONE_USER_SIZE, (const char *) name);
-#else
-    return 0;
-#endif
-}
-
 INTERNAL_HIDDEN int32_t name_mapping(void *p, size_t sz, const char *name) {
 #if NAMED_MAPPINGS && __ANDROID__
     return prctl(PR_SET_VMA, PR_SET_VMA_ANON_NAME, p, sz, name);
